Refactor to use the downloadSVG function in the profiles
